+++ src/commands/open-wiki.ts
@@ -5,19 +5,19 @@ import { Environment } from "../extension";
 import github from "../github";
 
 export default async function openWiki(env: Environment) {
-  // get user's github session
+  // Get user's github session
   const session = await vscode.authentication.getSession("github", ["repo"], {
     createIfNone: true,
   });
 
-  // get user's repos
+  // Get user's repos via session
   github.updateToken(session?.accessToken);
   let repos = await github.getUserRepos();
   repos = repos.filter((repo) => {
     return repo.has_wiki && !repo.archived;
   });
 
-  // prompt user to select a wiki
+  // Prompt user to select a wiki
   const repo = (
     await vscode.window.showQuickPick(
       repos.map(
@@ -43,7 +43,7 @@ export default async function openWiki(env: Environment) {
   const wikiSourcePath = path.join(wikiBasePath, "wiki");
   const wikiWorkspacePath = path.join(wikiBasePath, "wiki.code-workspace");
 
-  // check if wiki workspace already exists
+  // Check if the wiki workspace already exists for that repo
   if (fs.existsSync(wikiWorkspacePath)) {
     const newWorkspace = await vscode.window.showInformationMessage(
       "Wiki workspace already exists. Open a new workspace?",
@@ -51,16 +51,18 @@ export default async function openWiki(env: Environment) {
       "No"
     );
     if (newWorkspace !== "Yes") {
+      // Opens the existing wiki workspace
       vscode.commands.executeCommand(
         "vscode.openFolder",
         vscode.Uri.file(wikiWorkspacePath)
       );
       return;
     }
+    // Deletes the existing wiki workspace
     fs.rmSync(wikiBasePath, { recursive: true });
   }
 
-  // create wiki workspace
+  // Create a wiki workspace for that repo
   fs.mkdirSync(wikiSourcePath, { recursive: true });
   fs.writeFileSync(
     wikiWorkspacePath,
@@ -84,7 +86,7 @@ export default async function openWiki(env: Environment) {
     })
   );
 
-  // open wiki workspace
+  // Open wiki workspace
   vscode.commands.executeCommand(
     "vscode.openFolder",
     vscode.Uri.file(wikiWorkspacePath)