+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this extension will be documented in this file.
 
+## [1.1.1] - 2024-03-15
+
+The changes for this version is documented [here](https://github.com/dentolos19/git-wiki-editor/releases/tag/v1.1.1).
+
 ## [1.1.0] - 2024-03-05
 
 The changes for this version is documented [here](https://github.com/dentolos19/git-wiki-editor/releases/tag/v1.1.0).